the beginning of the revision history
Here we have both the data files and the R files that we are using to
parse them. There is an external reference to FitTest. That system call
relies on libraries provided elsewhere, and returns data that summarize
goodness of fit in a less arbitrary fashion than R2 slope and rmse.