added files for each component by year and full
each chemical constituent has its own file set up because they are not
all the same--chiefly the number of fingerprints in the calibration data
set requires some changes to how the year by year model is calibrated
and run. Sadly, most of these files are pretty similar, but I need to
look at them each one by one in order to see what is going on with each
constitutent.
Yay.