made a subsetting function
This makes it possible to change the subset windows more or less on the
fly--by specifying a vector of start times and a vector of stop times,
the subset function will trim down fingerprints etc, and return a list
object with time, fingerprints, and chemical data. very nuanced, I am
sure.