modelExecutionFunction
this acts as a wrapper for the plsr function, it. makes executing a loop
a little more easy. also added a snippet of text to write a jpeg file of
each chemograph overlaying a hydrograph. this will allow a savvy data
ponderer to look at how the various fits actually fit, and how well the
various summary statistics capture how well predictions fit the time
series.  I probably should have added a line or two to plot the points
used in the calibration file.
The modelPrototype file contains code that I have used to make a
comprehensive model run. It uses all available calibration points from
each lab analyte, as well as each fingerprint type to cycle through 3
through 9 components in the PLSR model. All of this is done first for
each year, and then for the full 3-4 year time window.
An output table is recorded, and those plots I mentioned above are spat
out into an output directory. joy of joys, it is data season.