allows skip-rows for pruning away bad data
had to make a few changes in the load data function to accept a vector
of rows to remove--it could be changed, i suppose, to specific POSIX
objects, but that isn't how I wrote it because i was lazy, and it is
less than super easy to get date and time's from a plot. As it was, I
had to use a combination of excel and R to identify points to exclude.

Scratch pad continues to accumulate code snippets that i work on before
integrating them into the main code body. it is not a place any
reasonable human other than me should ever look. ut1 and ut2 are the
same, and I will absorb them into the scratchpad file when i am brave
enough. (but don't look at them, unless you are me.)
Load data had a couple conditions added to it to allow loading of data.
I think that I need to convert some of these highly specific actions
(loading files) to be a bit more general so someone else has less
trouble integrating their own files. Something to think about.

Yours truly,
Kyle